Bump dependency and plugin versions. Fix warnings.
Note that there is an issue between immutables generated code and error
prone that needs to be readdressed when immutables is fixed.

+++ src/main/java/net/darkmist/chunks/Chunk.java
@@ -252,7 +252,7 @@ public final class Chunk extends AbstractList<Byte> implements Serializable, Com
 	@Override
 	public final Byte get(int off)
 	{	// pass directly down so we don't needlessly convert off from int to long to int
-		return Byte.valueOf((byte)(spi.getByte(off)));
+		return (byte)(spi.getByte(off));
 	}
 
 	/**
